+++ .npmignore
@@ -2,8 +2,9 @@ node_modules
 .env
 .vscode
 .git
-src/test
+test
 .prettierrc
 .npmignore
 .gitignore
 eslint.config.mjs
+package-lock.json
