--- index.js
@@ -1,5 +1,0 @@
-import skeleton from "./src/skeleton.js";
-
-export { TYPE } from "./src/consts/Type.js";
-export { JOIN } from "./src/consts/Join.js";
-export default skeleton;
