+++ lib/Parser.js
@@ -1,5 +1,5 @@
 import { wrapBacktick } from "./utils/backtick.js";
-import Validator from "./Validator.js";
+import Validator from "./utils/Validator.js";
 
 class Parser {
     static cols(cols) {
