+++ lib/QueryExecutor.js
