+++ lib/constants/ERROR_MESSAGE.js
