+++ lib/constants/JOIN.js
