+++ lib/constants/TYPE.js
