+++ lib/crud.js
@@ -46,7 +46,7 @@ export function remove({ from, where, orderBy, limit }) {
     return `DELETE ${[from, where, orderBy, limit].filter((e) => e).join(" ")}`;
 }
 
-export function insert({ cols, into, values, where }) {
+export function insert({ cols, into, values }) {
     into = `INTO ${Parser.into(into)}`; // 테이블 이름 변환
     values = `VALUES ${Parser.values(values)}`; // 값을 변환
     cols = cols ? `(${Parser.cols(cols)})` : ""; // cols 인자를 올바르게 처리
