+++ lib/index.js
@@ -1,0 +1,5 @@
+import skeleton from "./skeleton.js";
+
+export { default as TYPE } from "./constants/TYPE.js";
+export { default as JOIN } from "./constants/JOIN.js";
+export default skeleton;
