+++ lib/skeleton.js
@@ -1,6 +1,6 @@
-import { select, insert, remove, update } from "./CRUD.js";
-import { TYPE } from "./consts/Type.js";
-import { ErrorMessage } from "./consts/Error.js";
+import { select, insert, remove, update } from "./crud.js";
+import TYPE from "./constants/TYPE.js";
+import ERROR_MESSAGE from "./constants/ERROR_MESSAGE.js";
 import QueryExecutor from "./QueryExecutor.js";
 
 function literalize(sql) {
@@ -14,7 +14,7 @@ function literalize(sql) {
         case TYPE.UPDATE:
             return update(sql);
         default:
-            throw TypeError(ErrorMessage.type);
+            throw TypeError(ERROR_MESSAGE.TYPE);
     }
 }
