+++ lib/utils/Validator.js
@@ -1,5 +1,5 @@
-import ERROR_MESSAGE from "./constants/ERROR_MESSAGE.js";
-import { JOIN } from "./constants/JOIN.js";
+import ERROR_MESSAGE from "../constants/ERROR_MESSAGE.js";
+import JOIN from "../constants/JOIN.js";
 
 class TypeChecker {
     /**
@@ -83,7 +83,7 @@ class Validator {
     static checkValues(values) {
         // 2차원 배열인지 확인
         const isArrayValues = Array.isArray(values);
-        if (!isArrayValues) throw TypeError(ERROR_MESSAGE.values.twoDimensionArray);
+        if (!isArrayValues) throw TypeError(ERROR_MESSAGE.VALUES.TWO_DIMENSION_ARRAY);
         const isArrayValue = values.every((value) => Array.isArray(value));
         if (!isArrayValue) throw TypeError(ERROR_MESSAGE.VALUES.ARRAY);
         // 원소의 타입 체크
