+++ package.json
@@ -2,12 +2,11 @@
     "name": "devcourse_skeleton",
     "version": "0.0.7",
     "description": "Express + MySQL",
-    "main": "index.js",
+    "main": "lib/index.js",
     "scripts": {
-        "test": "node src/test/CRUD.test.js",
+        "test": "node src/test/index.test.js",
         "lint": "eslint"
     },
-    "author": "",
     "license": "MIT",
     "contributors": [
         "JaeHyeok Lee <ahhpc2012@gmail.com> (https://github.com/JaeHye0k)",
